refactored to basic less.css archetecture
