+++ src/components/common/BreadCrumbs.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Breadcrumb } from 'antd';
-import './BreadCrumbs.css';
+import '../../styles/css/BreadCrumbs.css';
 
 const breadCrumbNameMap = {
   '/layouttemplate': 'Layout Template',
