+++ src/components/pages/LayoutTemplate/RenderLayoutTemplate.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useLocation } from 'react-router-dom';
 import { Layout, Card } from 'antd';
-import './RenderLayoutTemplate.css';
+import '../../../styles/css/RenderLayoutTemplate.css';
 import NavHeader from '../../common/NavHeader';
 import BreadCrumbs from '../../common/BreadCrumbs';
 const { Header, Content, Footer } = Layout;
