+++ src/styles/css/RenderLayoutTemplate.css
@@ -1,0 +1,22 @@
+.RenderLayoutTemplate {
+  height: 100vh;
+}
+.layout {
+  height: 100%;
+}
+.Header {
+  background: white;
+  height: 8%;
+}
+.Content {
+  padding: 0 50px;
+}
+.site-layout-content {
+  height: 90%;
+}
+.Card {
+  height: 100%;
+}
+.Footer {
+  text-align: center;
+}
